Support negative scores
Scores can go negative. In the test custom game, when the team scores -1,
the API returns the max value of a 32-bit unsigned int. This commit adds a
converter that will cast the value from unsigned to signed when coming in
and back from signed to unsigned when going out.